twitter-data-analyzer/requirements.txt
tweepy>=4.14.0
duckdb>=0.10.0
google-generativeai>=0.3.0
typer>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
